==> 1. Logger_preparation.Rmd <==
# Logger data preparation

*Code reproducibility*

To allow a better reproducibility of this script, we recommend to use checkpoint() set at 2022-11-21. This allows to install and use the packages that were available and run at the time of making this script.

```{r}
#install.packages("checkpoint")
library(checkpoint)
checkpoint("2022-11-21")
```

Load packages:

```{r}
library(readxl)
library(lubridate)
library(dplyr)
library(ggplot2)
library(plyr)
library(tidyr)
library(reshape2)
library(naniar)
library(zoo)
library(ncdf4)
library(raster)
library(maptools)
library(checkpoint)
#library(GGally)
```

## Load all data:

### Site B (Arville):

```{r}
file.list <- list.files(path = "J:/TEMPERATURE DATA/ARVILLE", 
                        pattern = "*.xlsx", all.files = FALSE, full.names = TRUE, recursive = FALSE,
                        ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE)

df.list <- lapply(file.list, read_excel)
rm(file.list)
arville.temp <- do.call(rbind, df.list)

#oneyear.arville <- seq(ymd_hm('2020-2-1 00:00'),ymd_hm('2021-2-1 00:00'), by = '15 mins')
#arville.temp <- arville.temp[which(arville.temp$Date_time %in% oneyear.arville),]

arville.temp$site <- "Arville"
arville.temp$location <- "a"
arville.temp <- separate(arville.temp, Date_time, into = c("Date", "time"), sep = " (?=[^ ]+$)")

rm(df.list)
```

### Site G (Hochwald):

```{r}
file.list <- list.files(path = "J:/TEMPERATURE DATA/HOCHWALD/", 
                        pattern = "*.xlsx", all.files = FALSE, full.names = TRUE, recursive = FALSE,
                        ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE)

df.list <- lapply(file.list, read_excel)
rm(file.list)
hochwald.temp <- do.call(rbind, df.list)

#oneyear.hochwald <- seq(ymd_hm('2020-4-1 00:00'),ymd_hm('2021-4-1 00:00'), by = '15 mins')
#hochwald.temp <- hochwald.temp[which(hochwald.temp$Date_time %in% oneyear.hochwald),]

hochwald.temp$site <- "Hochwald"
hochwald.temp <- separate(hochwald.temp, Date_time, into = c("Date", "time"), sep = " (?=[^ ]+$)")

rm(df.list)
```

### Site SS (Vivarp):

```{r}
file.list <- list.files(path = "J:/TEMPERATURE DATA/VIVARP/", pattern = "*.csv", all.files = FALSE,
                        full.names = TRUE, recursive = TRUE,
                        ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE)

df.list <- lapply(file.list, read.csv)
rm(file.list)
vivarp.temp <- do.call(rbind, df.list)

# vivarp.temp <- dplyr::rename(vivarp.temp, 
#                       "Plot" = "ï..Plot")

vivarp.temp <- na.omit(vivarp.temp)

#Plot 16 used a different logger type that recorded temperatures every 15 min. Let's adjust that to only keeping temperatures at times being o'clock and half-past. Temperature measurements for plot 16 start at quarter-past, and record every 15 min (i.e. XX:15, XX:30, XX:45, XX:00 and so on). Therefore, to keep only temperatures recorded at o'clock and half-past, we'll just keep even rows. 

plot16 <- vivarp.temp[which(vivarp.temp$Plot == 16),]
plot16 <- plot16[seq(0, length(plot16$DateTime), 2),]

#Now remove such plots from the main dataset ( _temp.data_ ) and merge the fixed version with temperatures taken at :00 and :30

vivarp.temp <- vivarp.temp[-which(vivarp.temp$Plot == 16),]
vivarp.temp <- rbind(vivarp.temp, plot16)

rm(plot16)

#Select one year of data
#oneyear.vivarp <- seq(from = as.POSIXct("03/07/2020 00:00:00", format = "%d/%m/%Y %H:%M:%S"), to = as.POSIXct("03/07/2021 00:00:00", format = "%d/%m/%Y %H:%M:%S"), by = "30 min")

vivarp.temp$DateTime <- as.POSIXct(vivarp.temp$DateTime, format = "%d/%m/%Y %H:%M:%S")

#vivarp.temp <- vivarp.temp[which(vivarp.temp$DateTime %in% oneyear.vivarp),]

vivarp.temp$site <- "Vivarp"
vivarp.temp$location <- "a"
vivarp.temp <- separate(vivarp.temp, DateTime, into = c("Date", "time"), sep = " (?=[^ ]+$)")

rm(df.list)
```

Soil loggers in Vivarp were from a different model. We'll load them separately:

```{r}
file.list <- list.files(path = "J:/TEMPERATURE DATA/VIVARP/SOIL", pattern = "*.xlsx", all.files = FALSE,
                        full.names = TRUE, recursive = TRUE,
                        ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE)

df.list <- lapply(file.list, read_excel)
rm(file.list)
vivarp.temp.2 <- do.call(rbind, df.list)

vivarp.temp.2 <- na.omit(vivarp.temp.2)

vivarp.temp.2$Date_time <- as.POSIXct(vivarp.temp.2$Date_time, format = "%d/%m/%Y %H:%M:%S")

#oneyear.vivarp.s <- seq(from = as.POSIXct("21/10/2019 00:00:00", format = "%d/%m/%Y %H:%M:%S"), to = as.POSIXct("21/10/2020 00:00:00", format = "%d/%m/%Y %H:%M:%S"), by = "30 min")

#vivarp.temp.2 <- vivarp.temp.2[which(vivarp.temp.2$Date_time %in% oneyear.vivarp.s),]

vivarp.temp.2 <- separate(vivarp.temp.2, Date_time, into = c("Date", "time"), sep = " (?=[^ ]+$)")

rm(df.list)
```

Merge both Vivarp temperature objects:

```{r}
vivarp.temp.2$site <- "Vivarp"
vivarp.temp <- rbind(vivarp.temp, vivarp.temp.2)
rm(vivarp.temp.2)
```


### Site MS (Siljan):

```{r}
file.list <- list.files(path = "J:/TEMPERATURE DATA/SILJAN/", 
                        pattern = "*.xlsx", all.files = FALSE, full.names = TRUE,
                        recursive = FALSE, ignore.case = FALSE, include.dirs = FALSE,                         no.. = FALSE)

df.list <- lapply(file.list, read_excel)
rm(file.list)
siljan.temp <- do.call(rbind, df.list)

#oneyear.siljan <- seq(ymd_hm('2019-9-1 00:00'),ymd_hm('2020-9-1 00:00'), by = '15 mins')
#siljan.temp <- siljan.temp[which(siljan.temp$Date_time %in% oneyear.siljan),]

siljan.temp$site <- "Siljan"
siljan.temp <- separate(siljan.temp, Date_time, into = c("Date", "time"), sep = " (?=[^ ]+$)")

rm(df.list)
```

### Site NS (Sandsjö):

```{r}
file.list <- list.files(path = "J:/TEMPERATURE DATA/SANDSJO/", 
                        pattern = "*.xlsx", all.files = FALSE, full.names = TRUE,
                        recursive = FALSE, ignore.case = FALSE, include.dirs = FALSE,
                        no.. = FALSE)

df.list <- lapply(file.list, read_excel)
rm(file.list)
sandsjo.temp <- do.call(rbind, df.list)

#oneyear.sandsjo <- seq(ymd_hm('2019-9-2 00:00'),ymd_hm('2020-9-2 00:00'), by = '15 mins')
#sandsjo.temp <- sandsjo.temp[which(sandsjo.temp$Date_time %in% oneyear.sandsjo),]

sandsjo.temp$site <- "Sandsjo"
sandsjo.temp <- separate(sandsjo.temp, Date_time, into = c("Date", "time"), sep = " (?=[^ ]+$)")

rm(df.list)
```

## Check data completeness and remove problematic plots

### Site B (Arville):

```{r}
# In Arville, only air temperature was recorded. But we'll create a new variable (arville.temp.a) just for the sake of nomenclature consistence of this script

arville.temp.a <- arville.temp[which(arville.temp$location == "a"),]

# Select coldest and warmest months
coldmonth.arville.a <- seq(ymd_hm('2021-1-1 00:00'),ymd_hm('2021-1-31 00:00'), by = '15 mins')

warmmonth.arville.a <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

# create a month column
arville.temp.a$month <- month(arville.temp.a$Date)
arville.temp.a$Date <- as.POSIXct(arville.temp.a$Date)

# create an object for January and another one for July
arville.temp.a.jan <- arville.temp.a[which(arville.temp.a$Date %in% coldmonth.arville.a),]

arville.temp.a.jul <- arville.temp.a[which(arville.temp.a$Date %in% warmmonth.arville.a),]

arville.completeness.a.jan <- dcast(arville.temp.a.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

arville.completeness.a.jul <- dcast(arville.temp.a.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vis_miss(arville.completeness.a.jan, warn_large_data = FALSE)
vis_miss(arville.completeness.a.jul, warn_large_data = FALSE)
```

Loggers disappeared (i.e. all data missing) in the following plots (they are already missing from the dataset):
3, 17, 26, 32, 33


```{r}
arville.temp <- rbind(arville.temp.a.jan, arville.temp.a.jul)

rm(arville.temp.a)
rm(arville.temp.a.jan)
rm(arville.temp.a.jul)
rm(arville.completeness.a.jan)
rm(arville.completeness.a.jul)
rm(coldmonth.arville.a)
rm(warmmonth.arville.a)
```

### Site G (Hochwald):

```{r}
hochwald.temp$month <- month(hochwald.temp$Date)
# Separate soil and air logger data
hochwald.temp.a <- hochwald.temp[which(hochwald.temp$location == "a"),]

hochwald.temp.s <- hochwald.temp[which(hochwald.temp$location == "s"),]

# Select coldest and warmest months
coldmonth.hochwald.a <- seq(ymd_hm('2021-1-1 00:00'),ymd_hm('2021-1-31 00:00'), by = '15 mins')

warmmonth.hochwald.a <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

coldmonth.hochwald.s <- seq(ymd_hm('2021-1-1 00:00'),ymd_hm('2021-1-31 00:00'), by = '15 mins')

warmmonth.hochwald.s <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

hochwald.temp.a$Date <- as.POSIXct(hochwald.temp.a$Date)
hochwald.temp.s$Date <- as.POSIXct(hochwald.temp.s$Date)

# create an object for January and another one for July
hochwald.temp.a.jan <- hochwald.temp.a[which(hochwald.temp.a$Date %in% coldmonth.hochwald.a),]

hochwald.temp.a.jul <- hochwald.temp.a[which(hochwald.temp.a$Date %in% warmmonth.hochwald.a),]

hochwald.temp.s.jan <- hochwald.temp.s[which(hochwald.temp.s$Date %in% coldmonth.hochwald.s),]

hochwald.temp.s.jul <- hochwald.temp.s[which(hochwald.temp.s$Date %in% warmmonth.hochwald.s),]

hochwald.completeness.a.jan <- dcast(hochwald.temp.a.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

hochwald.completeness.a.jul <- dcast(hochwald.temp.a.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

hochwald.completeness.s.jan <- dcast(hochwald.temp.s.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

hochwald.completeness.s.jul <- dcast(hochwald.temp.s.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vis_miss(hochwald.completeness.a.jan, warn_large_data = FALSE)
vis_miss(hochwald.completeness.a.jul, warn_large_data = FALSE)
vis_miss(hochwald.completeness.s.jan, warn_large_data = FALSE)
vis_miss(hochwald.completeness.s.jul, warn_large_data = FALSE)

rm(hochwald.completeness.a.jan)
rm(hochwald.completeness.a.jul)
rm(hochwald.completeness.s.jan)
rm(hochwald.completeness.s.jul)
rm(coldmonth.hochwald.a)
rm(coldmonth.hochwald.s)
rm(warmmonth.hochwald.a)
rm(warmmonth.hochwald.s)

rm(hochwald.temp)
rm(hochwald.temp.a)
rm(hochwald.temp.s)
```

Loggers that failed (all data missing):

_Air logger in January:_
- 6
- 15
- 34
- 39

_Air logger in July:_
- 15
- 39

_Soil logger in January:_
- 29

_Soil logger un July:_
- 29

These data are already absent in the subsetted dataset.
There are no loggers with data partially missing.

```{r}
hochwald.temp <- rbind(hochwald.temp.a.jan, hochwald.temp.a.jul, hochwald.temp.s.jan, hochwald.temp.s.jul)

rm(hochwald.temp.a.jan)
rm(hochwald.temp.a.jul)
rm(hochwald.temp.s.jan)
rm(hochwald.temp.s.jul)
```

### Site SS (Vivarp):

```{r}
vivarp.temp$month <- month(vivarp.temp$Date)

# Separate soil and air logger data
vivarp.temp.a <- vivarp.temp[which(vivarp.temp$location == "a"),]

vivarp.temp.s <- vivarp.temp[which(vivarp.temp$location == "s"),]

rm(vivarp.temp)

# Select coldest and warmest months
coldmonth.vivarp.a <- seq(ymd_hm('2021-1-1 00:00'),ymd_hm('2021-1-31 00:00'), by = '15 mins')

warmmonth.vivarp.a <- seq(ymd_hm('2021-7-1 00:00'),ymd_hm('2021-7-31 00:00'), by = '15 mins')

coldmonth.vivarp.s <- seq(ymd_hm('2020-1-1 00:00'),ymd_hm('2020-1-31 00:00'), by = '15 mins')

warmmonth.vivarp.s <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

vivarp.temp.a$Date <- as.POSIXct(vivarp.temp.a$Date)
vivarp.temp.s$Date <- as.POSIXct(vivarp.temp.s$Date)

# create an object for January and another one for July
vivarp.temp.a.jan <- vivarp.temp.a[which(vivarp.temp.a$Date %in% coldmonth.vivarp.a),]

vivarp.temp.a.jul <- vivarp.temp.a[which(vivarp.temp.a$Date %in% warmmonth.vivarp.a),]

vivarp.temp.s.jan <- vivarp.temp.s[which(vivarp.temp.s$Date %in% coldmonth.vivarp.s),]

vivarp.temp.s.jul <- vivarp.temp.s[which(vivarp.temp.s$Date %in% warmmonth.vivarp.s),]

# Data completeness
vivarp.completeness.a.jan <- dcast(vivarp.temp.a.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vivarp.completeness.a.jul <- dcast(vivarp.temp.a.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vivarp.completeness.s.jan <- dcast(vivarp.temp.s.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vivarp.completeness.s.jul <- dcast(vivarp.temp.s.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vis_miss(vivarp.completeness.a.jan, warn_large_data = FALSE)
vis_miss(vivarp.completeness.a.jul, warn_large_data = FALSE)
vis_miss(vivarp.completeness.s.jan, warn_large_data = FALSE)
vis_miss(vivarp.completeness.s.jul, warn_large_data = FALSE)

rm(vivarp.completeness.a.jan)
rm(vivarp.completeness.a.jul)
rm(vivarp.completeness.s.jan)
rm(vivarp.completeness.s.jul)
rm(coldmonth.vivarp.a)
rm(coldmonth.vivarp.s)
rm(warmmonth.vivarp.a)
rm(warmmonth.vivarp.s)
```

Loggers that failed (all data missing):

_Air logger in January:_
- 30

_Air logger in July:_
- Nothing missing

_Soil logger in January:_
- 18
- 27

_Soil logger un July:_
- Nothing missing

These data are already absent in the subsetted dataset.
There are no loggers with data partially missing.

```{r}
vivarp.temp <- rbind(vivarp.temp.a.jan, vivarp.temp.a.jul, vivarp.temp.s.jan, vivarp.temp.s.jul)

rm(vivarp.temp.a.jan)
rm(vivarp.temp.a.jul)
rm(vivarp.temp.s.jan)
rm(vivarp.temp.s.jul)
rm(vivarp.temp.a)
rm(vivarp.temp.s)
```

### Site MS (Siljan):

```{r}
siljan.temp$month <- month(siljan.temp$Date)

# Separate soil and air logger data
siljan.temp.a <- siljan.temp[which(siljan.temp$location == "a"),]

siljan.temp.s <- siljan.temp[which(siljan.temp$location == "s"),]
rm(siljan.temp)

# Select coldest and warmest months
coldmonth.siljan.a <- seq(ymd_hm('2021-1-1 00:00'),ymd_hm('2021-1-31 00:00'), by = '15 mins')

warmmonth.siljan.a <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

coldmonth.siljan.s <- seq(ymd_hm('2021-1-1 00:00'),ymd_hm('2021-1-31 00:00'), by = '15 mins')

warmmonth.siljan.s <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

siljan.temp.a$Date <- as.POSIXct(siljan.temp.a$Date)
siljan.temp.s$Date <- as.POSIXct(siljan.temp.s$Date)

# create an object for January and another one for July
siljan.temp.a.jan <- siljan.temp.a[which(siljan.temp.a$Date %in% coldmonth.siljan.a),]

siljan.temp.a.jul <- siljan.temp.a[which(siljan.temp.a$Date %in% warmmonth.siljan.a),]

siljan.temp.s.jan <- siljan.temp.s[which(siljan.temp.s$Date %in% coldmonth.siljan.s),]

siljan.temp.s.jul <- siljan.temp.s[which(siljan.temp.s$Date %in% warmmonth.siljan.s),]

# Data completeness
siljan.completeness.a.jan <- dcast(siljan.temp.a.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

siljan.completeness.a.jul <- dcast(siljan.temp.a.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

siljan.completeness.s.jan <- dcast(siljan.temp.s.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

siljan.completeness.s.jul <- dcast(siljan.temp.s.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vis_miss(siljan.completeness.a.jan, warn_large_data = FALSE)
vis_miss(siljan.completeness.a.jul, warn_large_data = FALSE)
vis_miss(siljan.completeness.s.jan, warn_large_data = FALSE)
vis_miss(siljan.completeness.s.jul, warn_large_data = FALSE)

rm(siljan.completeness.a.jan)
rm(siljan.completeness.a.jul)
rm(siljan.completeness.s.jan)
rm(siljan.completeness.s.jul)
```

Loggers that failed:

_Air logger in January:_
- 12 (86.67% data mising)
- 18 (100% data missing)

_Air logger in July:_
- 12 (100% data missing)
- 14 (100% data missing)
- 18 (100% data missing)
- 21 (100% data missing)
- 30 (100% data missing)

_Soil logger in January:_
- No data missing

_Soil logger un July:_
- 2 (100% data missing)
- 13 (100% data missing)
- 33 (100% data missing)

Loggers with 100% of data missing are already absent in the subsetted dataset.

Air logger 12 is missing 86.67% of the data in January.

```{r}
# Remove air logger 12 in the January dataset
siljan.temp.a.jan <- siljan.temp.a.jan[-which(siljan.temp.a.jan$Plot == 12),]

siljan.temp <- rbind(siljan.temp.a.jan, siljan.temp.a.jul, siljan.temp.s.jan, siljan.temp.s.jul)

rm(siljan.temp.a.jan)
rm(siljan.temp.a.jul)
rm(siljan.temp.s.jan)
rm(siljan.temp.s.jul)
rm(siljan.temp.a)
rm(siljan.temp.s)
rm(coldmonth.siljan.a)
rm(coldmonth.siljan.s)
rm(warmmonth.siljan.a)
rm(warmmonth.siljan.s)
```


### Site NS (Sandsjö):

```{r}
sandsjo.temp$month <- month(sandsjo.temp$Date)

# Separate soil and air logger data
sandsjo.temp.a <- sandsjo.temp[which(sandsjo.temp$location == "a"),]

sandsjo.temp.s <- sandsjo.temp[which(sandsjo.temp$location == "s"),]
rm(sandsjo.temp)

# Select coldest and warmest months
coldmonth.sandsjo.a <- seq(ymd_hm('2020-1-1 00:00'),ymd_hm('2020-1-31 00:00'), by = '15 mins')

warmmonth.sandsjo.a <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

coldmonth.sandsjo.s <- seq(ymd_hm('2020-1-1 00:00'),ymd_hm('2020-1-31 00:00'), by = '15 mins')

warmmonth.sandsjo.s <- seq(ymd_hm('2020-7-1 00:00'),ymd_hm('2020-7-31 00:00'), by = '15 mins')

sandsjo.temp.a$Date <- as.POSIXct(sandsjo.temp.a$Date)
sandsjo.temp.s$Date <- as.POSIXct(sandsjo.temp.s$Date)

# create an object for January and another one for July
sandsjo.temp.a.jan <- sandsjo.temp.a[which(sandsjo.temp.a$Date %in% coldmonth.sandsjo.a),]

sandsjo.temp.a.jul <- sandsjo.temp.a[which(sandsjo.temp.a$Date %in% warmmonth.sandsjo.a),]

sandsjo.temp.s.jan <- sandsjo.temp.s[which(sandsjo.temp.s$Date %in% coldmonth.sandsjo.s),]

sandsjo.temp.s.jul <- sandsjo.temp.s[which(sandsjo.temp.s$Date %in% warmmonth.sandsjo.s),]

# Data completeness
sandsjo.completeness.a.jan <- dcast(sandsjo.temp.a.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

sandsjo.completeness.a.jul <- dcast(sandsjo.temp.a.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

sandsjo.completeness.s.jan <- dcast(sandsjo.temp.s.jan, Date~Plot, fun.aggregate = mean, value.var = "Temp")

sandsjo.completeness.s.jul <- dcast(sandsjo.temp.s.jul, Date~Plot, fun.aggregate = mean, value.var = "Temp")

vis_miss(sandsjo.completeness.a.jan, warn_large_data = FALSE)
vis_miss(sandsjo.completeness.a.jul, warn_large_data = FALSE)
vis_miss(sandsjo.completeness.s.jan, warn_large_data = FALSE)
vis_miss(sandsjo.completeness.s.jul, warn_large_data = FALSE)
```

Loggers that failed:

_Air logger in January:_
- 37 (100% data missing)

_Air logger in July:_
- 20 (63.33% missing)
- 37 (3.33% missing)

_Soil logger in January:_
- 36 (100% data missing)

_Soil logger un July:_
- 31 (100% data missing)
- 36 (100% data missing)

Loggers with 100% of data missing are already absent in the subsetted dataset.

Air logger 20 is missing 63.33% of the data in July, and air logger 37, 3.33%.


```{r}
# Remove air logger 20 in the July dataset
sandsjo.temp.a.jul <- sandsjo.temp.a.jul[-which(sandsjo.temp.a.jul$Plot == 20),]

# Data filling: as only 3.33% of data is missing from air logger 37 in July, we will fill this data gap with the average temperature obtained from the rest of the loggers in the same dates and time in the same forest stand.

# Get Dates with missing values for plot 35a in Vivarp
missing.sandsjo.37a.jul <- sandsjo.completeness.a.jul[is.na(sandsjo.completeness.a.jul$`37`),]$Date

# Get all the raw data from all loggers but 35 a for these dates
tofill.sandsjo.37a.jul <- sandsjo.temp.a.jul[which(sandsjo.temp.a.jul$Date %in% missing.sandsjo.37a.jul),]

# Calculate the mean across all plots per date in the missing dates
filled.sandsjo.37a.jul <- aggregate(list(Temp = tofill.sandsjo.37a.jul$Temp), by = list(Date = tofill.sandsjo.37a.jul$Date, time = tofill.sandsjo.37a.jul$time), FUN = mean)

# Create extra variables before merging to vivarp.temp.a
filled.sandsjo.37a.jul$Plot <- 37
filled.sandsjo.37a.jul$time <- NA
filled.sandsjo.37a.jul$site <- "Sandsjo"
filled.sandsjo.37a.jul$location <- "a"
filled.sandsjo.37a.jul$month <- month(filled.sandsjo.37a.jul$Date)

sandsjo.temp.a.jul <- rbind(sandsjo.temp.a.jul, filled.sandsjo.37a.jul)


rm(missing.sandsjo.37a.jul)
rm(tofill.sandsjo.37a.jul)
rm(filled.sandsjo.37a.jul)

sandsjo.temp <- rbind(sandsjo.temp.a.jan, sandsjo.temp.a.jul, sandsjo.temp.s.jan, sandsjo.temp.s.jul)

rm(sandsjo.temp.a.jan)
rm(sandsjo.temp.a.jul)
rm(sandsjo.temp.s.jan)
rm(sandsjo.temp.s.jul)

rm(coldmonth.sandsjo.a)
rm(coldmonth.sandsjo.s)
rm(warmmonth.sandsjo.a)
rm(warmmonth.sandsjo.s)

rm(sandsjo.completeness.a.jan)
rm(sandsjo.completeness.a.jul)
rm(sandsjo.completeness.s.jan)
rm(sandsjo.completeness.s.jul)
rm(sandsjo.temp.a)
rm(sandsjo.temp.s)
```

Merge and export as .csv for script 2.5

```{r}
extract <- rbind(arville.temp, hochwald.temp, vivarp.temp, siljan.temp, sandsjo.temp)

write.csv(extract, file = "./DATASETS/rawtemp_min.csv", row.names = FALSE, quote = FALSE)

rm(extract)
```

## Daily temperature variable calculation

Calculate mean and sd values per day and logger before merging:

```{r}
arville.temp <- do.call(data.frame, aggregate(Temp ~  Date + Plot + location + site, data = arville.temp, FUN = function(x) c(mean = mean(x), min = min(x), max = max(x), sd = sd(x))))

hochwald.temp <- do.call(data.frame, aggregate(Temp ~  Date + Plot + location + site, data = hochwald.temp, FUN = function(x) c(mean = mean(x), min = min(x), max = max(x), sd = sd(x))))

vivarp.temp <- do.call(data.frame, aggregate(Temp ~  Date + Plot + location + site, data = vivarp.temp, FUN = function(x) c(mean = mean(x), min = min(x), max = max(x), sd = sd(x))))

siljan.temp <- do.call(data.frame, aggregate(Temp ~  Date + Plot + location + site, data = siljan.temp, FUN = function(x) c(mean = mean(x), min = min(x), max = max(x), sd = sd(x))))

sandsjo.temp <- do.call(data.frame, aggregate(Temp ~  Date + Plot + location + site, data = sandsjo.temp, FUN = function(x) c(mean = mean(x), min = min(x), max = max(x), sd = sd(x))))
```

Export the dataset:

```{r}
temp.data <- rbind(arville.temp, hochwald.temp, sandsjo.temp, siljan.temp, vivarp.temp)

write.csv(temp.data, file = "./DATASETS/logger_temp.csv", row.names = FALSE, quote = FALSE)

rm(list = ls())
```
