==> 2.5 Macro vs Microclimate.Rmd <==
# Macro vs. Microclimate

*Code reproducibility*

To allow a better reproducibility of this script, we recommend to use checkpoint() set at 2022-11-21. This allows to install and use the packages that were available and run at the time of making this script.

```{r}
#install.packages("checkpoint")
library(checkpoint)
checkpoint("2022-11-21")
```

Load packages:

```{r}
library(lubridate)
library(ggplot2)
```

First, we need to load all data:

```{r}
logger.temp <- read.csv("./DATASETS/logger_temp.csv", header = TRUE)

ERA5.temp <- read.csv("./DATASETS/ERA5_temp.csv", header = TRUE)

mcera5.temp <- read.csv("./DATASETS/mcera5.csv", header = TRUE)

weatherstation.temp <- read.csv("./DATASETS/weatherstations_merged.csv", header = TRUE)
```

And we merge everything (weather station data, ERA5, mcera5 and our loggers) by site and date:

```{r}
temp <- merge(merge(merge(logger.temp, ERA5.temp, by = c("site", "Date"), all.x = TRUE), weatherstation.temp, by = c("site", "Date"), all.x = TRUE), mcera5.temp, by = c("site", "Date"), all.x = TRUE)

rm(logger.temp)
rm(ERA5.temp)
rm(weatherstation.temp)
rm(mcera5.temp)
```

Format:

```{r}
temp <- temp[which(temp$location == "a"),]
temp$month <- month(temp$Date)

arville <- temp[which(temp$site == "Arville"),]
hochwald <- temp[which(temp$site == "Hochwald"),]
vivarp <- temp[which(temp$site == "Vivarp"),]
siljan <- temp[which(temp$site == "Siljan"),]
sandsjo <- temp[which(temp$site == "Sandsjo"),]
```

Figure:

Arville:

```{r}
arville.week.cold <- arville[which(arville$month == 1),]

arville.coldplot <- ggplot(arville.week.cold[which(arville.week.cold$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 5),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 11),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 12),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 30),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 34),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 38),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.cold[which(arville.week.cold$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group = 1), colour = "#335c67", lwd = 1) + 
   ggtitle("Belgium coldest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("January 2021") +
    ylim(-23, 8)

  
#rm(arville.week.cold)

# Same with warmest month (july):

arville.week.warm <- arville[which(arville$month == 7),]

arville.warmplot <- 
  ggplot(arville.week.warm[which(arville.week.warm$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 5),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 11),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 12),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 30),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 34),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 38),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= arville.week.warm[which(arville.week.warm$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group = 1), colour = "#335c67", lwd = 1) +
   ggtitle("Belgium warmest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("July 2020") +
      ylim(7, 28)


  
rm(arville)
#rm(arville.week.warm)
```

Hochwald:

```{r}
hochwald.week.cold <- hochwald[which(hochwald$month == 1),]

hochwald.coldplot <- 
  ggplot(hochwald.week.cold[which(hochwald.week.cold$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 5),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 11),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 12),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 17),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 30),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 33),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 38),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.cold[which(hochwald.week.cold$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("Germany coldest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("January 2021") +
      ylim(-23, 8)


  
#rm(hochwald.week.cold)

# Same with warmest month (july):

hochwald.week.warm <- hochwald[which(hochwald$month == 7),]

hochwald.warmplot <- 
  ggplot(hochwald.week.warm[which(hochwald.week.warm$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 5),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 11),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 12),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 17),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 30),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 33),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 38),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= hochwald.week.warm[which(hochwald.week.warm$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("Germany warmest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("July 2020") +
      ylim(7, 28)



rm(hochwald)
#rm(hochwald.week.warm)
```

Vivarp:

```{r}
vivarp.week.cold <- vivarp[which(vivarp$mont == 1),]

vivarp.coldplot <- ggplot(vivarp.week.cold[which(vivarp.week.cold$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 5),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 11),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 12),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 17),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 33),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 34),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 38),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.cold[which(vivarp.week.cold$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("South Sweden coldest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("January 2021") +
      ylim(-23, 8)


  
#rm(vivarp.week.cold)

# Same with warmest month (july):

vivarp.week.warm <- vivarp[which(vivarp$month == 7),]

vivarp.warmplot <- 
  ggplot(vivarp.week.warm[which(vivarp.week.warm$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 2),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 5),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 11),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 12),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 17),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 25),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 27),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 28),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 31),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 33),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 34),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 38),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= vivarp.week.warm[which(vivarp.week.warm$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("South Sweden warmest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("July 2021") +
      ylim(7, 28)



rm(vivarp)
#rm(vivarp.week.warm)
```

Siljan:

```{r}
siljan.week.cold <- siljan[which(siljan$month == 1),]

siljan.coldplot <- 
  ggplot(siljan.week.cold[which(siljan.week.cold$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 5),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 11),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 17),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 33),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 34),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 38),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.cold[which(siljan.week.cold$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("Mid Sweden coldest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("January 2021") +
      ylim(-23, 8)


  
#rm(siljan.week.cold)

# Same with warmest month (july):

siljan.week.warm <- siljan[which(siljan$month == 7),]

siljan.warmplot <- 
  ggplot(siljan.week.warm[which(siljan.week.warm$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 2),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 5),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 11),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 17),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 20),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 25),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 27),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 28),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 29),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 31),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 33),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 34),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 37),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 38),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= siljan.week.warm[which(siljan.week.warm$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("Mid Sweden warmest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("July 2020") +
      ylim(7, 28)



rm(siljan)
#rm(siljan.week.warm)
```

Sandsjö:

```{r}
sandsjo.week.cold <- sandsjo[which(sandsjo$month == 1),]

sandsjo.coldplot <- 
  ggplot(sandsjo.week.cold[which(sandsjo.week.cold$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 2),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 5),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 11),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 12),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 13),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 17),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 19),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 25),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 27),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 28),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 29),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 30),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 31),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 33),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 34),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 38),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.cold[which(sandsjo.week.cold$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("North Sweden coldest month") +
    theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("January 2020") +
      ylim(-23, 8)



#rm(sandsjo.week.cold)

# Same with warmest month (july):

sandsjo.week.warm <- sandsjo[which(sandsjo$month == 7),]

sandsjo.warmplot <- 
  ggplot(sandsjo.week.warm[which(sandsjo.week.warm$Plot == 1),], aes(x = Date)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 1),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 2),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 3),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 4),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 5),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 6),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 7),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 8),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 9),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 10),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 11),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 12),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 13),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 14),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 15),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 16),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 17),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 18),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 19),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 21),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 22),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 23),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 24),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 25),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 26),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 27),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 28),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 29),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 30),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 31),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 32),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 33),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 34),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 35),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 36),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 38),]$Temp.mean, group = 1)) +
    geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 39),]$Temp.mean, group = 1)) +
  geom_line(color = "gray45", alpha = 0.5, lwd = 1, aes(y= sandsjo.week.warm[which(sandsjo.week.warm$Plot == 40),]$Temp.mean, group = 1)) + 
  geom_line(aes(y = era5_meanair, group = 1), colour = "#e09f3e", lwd = 1) +
    geom_line(aes(y = meanair_mcera5, group = 1), colour = "#9e2a2b", lwd = 1) +
    geom_line(aes(y = station_meanair, group =1), colour = "#335c67", lwd = 1) +
  ggtitle("North Sweden warmest month") +
  theme(axis.text.x = element_blank()) +
  ylab("Temperature (°C)") + 
  xlab("July 2020") +
      ylim(7, 28)


rm(sandsjo)
#rm(sandsjo.week.warm)
rm(temp)
```

Merge all plots:

```{r}
ggarrange(
sandsjo.coldplot,
sandsjo.warmplot,
siljan.coldplot,
siljan.warmplot,
vivarp.coldplot,
vivarp.warmplot,
hochwald.coldplot,
hochwald.warmplot,
arville.coldplot,
arville.warmplot,
ncol = 2,
nrow = 5,
common.legend = TRUE, 
legend = "bottom")
```
