==> 4. Tree_data_script.Rmd <==
# Overstory calculations

*Code reproducibility*

To allow a better reproducibility of this script, we recommend to use checkpoint() set at 2022-11-21. This allows to install and use the packages that were available and run at the time of making this script.

```{r}
#install.packages("checkpoint")
library(checkpoint)
checkpoint("2022-11-21")
```

Load packages:

```{r}
library(readxl)
library(dplyr)
library(ggplot2)
```


# Load tree data:

```{r}
file.list <- list.files(path = "J:/TREE DATA/", 
                        pattern = "*.xlsx", all.files = FALSE, full.names = TRUE, recursive = FALSE,
                        ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE)

df.list <- lapply(file.list, read_excel)

rm(file.list)
tree.data <- do.call(rbind, df.list)
rm(df.list)

tree.data <- tree.data[,-which(colnames(tree.data) == "Angle")]

tree.data$Distance <- as.numeric(tree.data$Distance)

tree.data <- na.omit(tree.data)
```

## Data cleaning

First, let's calculate the real distance from the hypothetical center of each tree (i.e. diameter/2) and the center of the plot. For this, we will add half the diameter of each tree to its recorded distance. Then, since we are only interested in trees within a 10 m radius, we will remove all trees with a "real distance" larger than 10m.

```{r}
tree.data$DBH <- as.numeric(tree.data$DBH)
tree.data$Distance <- as.numeric(tree.data$Distance)
tree.data$real.dist <- tree.data$Distance - tree.data$DBH/2000 #DBH is recorded in mm
```
We are also just interested in trees above 5cm of DBH. DBH is recorded in mm, so we will filter everything from 50 mm up.

```{r}
tree.data <- tree.data[which(tree.data$DBH >= 50),]
```

## Calculate basal area

We are interested in calculating how much basal area per site and tree species here was. For this, we will duplicate our main dataset.

```{r}
tree.data.sp <- tree.data

tree.data.sp$DBH <- as.numeric(tree.data.sp$DBH)
tree.data.sp$Distance <- as.numeric(tree.data.sp$Distance)
tree.data.sp$real.dist <- as.numeric(tree.data.sp$real.dist)

# First Arville, as we used 20 m radius to measure trees and there were only 38 plots instead of 40
tree.data.sp.arville <- tree.data.sp %>%
  filter((Site %in% "Arville")) %>%
  group_by(Site, Species) %>%
  summarise(
Basal_Area_sp = (sum((pi*(DBH[real.dist <= 20]/20)^2)/10000)/(38*pi*20^2))*10000)

tree.data.sp.arville$BA_total <- sum(tree.data.sp.arville$Basal_Area_sp) # add total BA

# In Hochwald we have 39 plots instead of 40. We'll treat it separately:
tree.data.sp.hoch <- tree.data.sp %>%
  filter((Site %in% "Hochwald")) %>%
  group_by(Site, Species) %>%
  summarise(
Basal_Area_sp = (sum((pi*(DBH[real.dist <= 10]/20)^2)/10000)/(39*pi*10^2))*10000)

tree.data.sp.hoch$BA_total <- sum(tree.data.sp.hoch$Basal_Area_sp) # add total BA

# Then other sites (with 40 plots):
tree.data.sp.noarv <- tree.data.sp %>%
  filter(!(Site %in% c("Arville", "Hochwald"))) %>%
  group_by(Site, Species) %>%
  summarise(
Basal_Area_sp = (sum((pi*(DBH[real.dist <= 10]/20)^2)/10000)/(40*pi*10^2))*10000)

# Add total BA per site:
tree.data.sp.noarv$BA_total <- 0 # empty variable

tree.data.sp.noarv[which(tree.data.sp.noarv$Site == "Vivarp"),]$BA_total <- sum(tree.data.sp.noarv[which(tree.data.sp.noarv$Site == "Vivarp"),]$Basal_Area_sp)

tree.data.sp.noarv[which(tree.data.sp.noarv$Site == "Siljan"),]$BA_total <- sum(tree.data.sp.noarv[which(tree.data.sp.noarv$Site == "Siljan"),]$Basal_Area_sp)

tree.data.sp.noarv[which(tree.data.sp.noarv$Site == "Sandsjo"),]$BA_total <- sum(tree.data.sp.noarv[which(tree.data.sp.noarv$Site == "Sandsjo"),]$Basal_Area_sp)

# Merge:

tree.data.sp <- rbind(tree.data.sp.arville, tree.data.sp.hoch, tree.data.sp.noarv)

rm(tree.data.sp.arville)
rm(tree.data.sp.hoch)
rm(tree.data.sp.noarv)

# Calculate % of BA per species and site:
tree.data.sp$perc_BA_sp <- round((tree.data.sp$Basal_Area_sp/tree.data.sp$BA_total)*100,2)

tree.data.sp

rm(tree.data.sp)
```

As we measured trees in a 10 m radius everywhere but in Arville, where we used 20 m radius, we will run this in two separate instructions to create variables for different radi at 1 m intervals. This way we will not calculate BA calues for radi between 11 and 20 for sites that are not Arville.

```{r}
#detach(package:plyr) #in case it has been loaded previously

tree.data$DBH <- as.numeric(tree.data$DBH)
tree.data$Distance <- as.numeric(tree.data$Distance)
tree.data$real.dist <- as.numeric(tree.data$real.dist)

tree.data.plot.arville <- tree.data %>%
  filter((Site %in% "Arville")) %>%
  group_by(Site, Plot) %>%
  summarise(
Basal_Area_20m = (sum((pi*(DBH[real.dist <= 20]/20)^2)/10000)/(pi*20^2))*10000,
Basal_Area_19m = (sum((pi*(DBH[real.dist <= 19]/20)^2)/10000)/(pi*19^2))*10000,
Basal_Area_18m = (sum((pi*(DBH[real.dist <= 18]/20)^2)/10000)/(pi*18^2))*10000,
Basal_Area_17m = (sum((pi*(DBH[real.dist <= 17]/20)^2)/10000)/(pi*17^2))*10000,
Basal_Area_16m = (sum((pi*(DBH[real.dist <= 16]/20)^2)/10000)/(pi*16^2))*10000,
Basal_Area_15m = (sum((pi*(DBH[real.dist <= 15]/20)^2)/10000)/(pi*15^2))*10000,
Basal_Area_14m = (sum((pi*(DBH[real.dist <= 14]/20)^2)/10000)/(pi*14^2))*10000,
Basal_Area_13m = (sum((pi*(DBH[real.dist <= 13]/20)^2)/10000)/(pi*13^2))*10000,
Basal_Area_12m = (sum((pi*(DBH[real.dist <= 12]/20)^2)/10000)/(pi*12^2))*10000,
Basal_Area_11m = (sum((pi*(DBH[real.dist <= 11]/20)^2)/10000)/(pi*11^2))*10000,
Basal_Area_10m = (sum((pi*(DBH[real.dist <= 10]/20)^2)/10000)/(pi*10^2))*10000,
Basal_Area_09m = (sum((pi*(DBH[real.dist <= 9]/20)^2)/10000)/(pi*9^2))*10000,
Basal_Area_08m = (sum((pi*(DBH[real.dist <= 8]/20)^2)/10000)/(pi*8^2))*10000,
Basal_Area_07m = (sum((pi*(DBH[real.dist <= 7]/20)^2)/10000)/(pi*7^2))*10000,
Basal_Area_06m = (sum((pi*(DBH[real.dist <= 6]/20)^2)/10000)/(pi*6^2))*10000,
Basal_Area_05m = (sum((pi*(DBH[real.dist <= 5]/20)^2)/10000)/(pi*5^2))*10000,
Basal_Area_04m = (sum((pi*(DBH[real.dist <= 4]/20)^2)/10000)/(pi*4^2))*10000)

tree.data.plot.notarville <- tree.data %>%
  filter(!(Site %in% "Arville")) %>%
  group_by(Site, Plot) %>%
  summarise(
Basal_Area_10m = (sum((pi*(DBH[real.dist <= 10]/20)^2)/10000)/(pi*10^2))*10000,
Basal_Area_09m = (sum((pi*(DBH[real.dist <= 9]/20)^2)/10000)/(pi*9^2))*10000,
Basal_Area_08m = (sum((pi*(DBH[real.dist <= 8]/20)^2)/10000)/(pi*8^2))*10000,
Basal_Area_07m = (sum((pi*(DBH[real.dist <= 7]/20)^2)/10000)/(pi*7^2))*10000,
Basal_Area_06m = (sum((pi*(DBH[real.dist <= 6]/20)^2)/10000)/(pi*6^2))*10000,
Basal_Area_05m = (sum((pi*(DBH[real.dist <= 5]/20)^2)/10000)/(pi*5^2))*10000,
Basal_Area_04m = (sum((pi*(DBH[real.dist <= 4]/20)^2)/10000)/(pi*4^2))*10000)

# As we did not measure trees between 10 and 20 m radi for places other than Arville, we will copy the BA value at 10 m radi to all other radi for these places. This way, when using the "Basal_Area_20m" variable we will be using all our available data (i.e BA calculated at 20 m in Arville and at 10 in all the other sites)

tree.data.plot.notarville$Basal_Area_20m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_19m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_18m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_17m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_16m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_15m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_14m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_13m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_12m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot.notarville$Basal_Area_11m <- tree.data.plot.notarville$Basal_Area_10m

tree.data.plot <- rbind (tree.data.plot.arville, tree.data.plot.notarville)

rm(tree.data.plot.arville)
rm(tree.data.plot.notarville)
```

# Percentage of broad-leaved species

First, we will create a new variable stating whether a tree species is broad-leaved or narrow-leaved.

```{r}
levels(as.factor(tree.data$Species))
```

A = Alder
B = Birch
BE = Beech
CB = Carpinus betulus
DF = Douglas fir
E = Elder
L = Larch
O = Oak
P = Pine
R = Rowan
S = Spruce
Salix = Salix
SITKA = Sitka

```{r}
tree.data$leaves <- ifelse(tree.data$Species %in% c("P", "S", "DF", "L", "SITKA"),"Narrow","Broad")

broad.leaved <- tree.data[which(tree.data$leaves == "Broad"),] 

tree.data.plot2 <- broad.leaved %>%
  filter((Site %in% "Arville")) %>%
  group_by(Site, Plot) %>%
  summarise(
Broad_BA_20m = (sum((pi*(DBH[real.dist <= 20]/20)^2)/10000)/(pi*20^2))*10000,
Broad_BA_19m = (sum((pi*(DBH[real.dist <= 19]/20)^2)/10000)/(pi*19^2))*10000,
Broad_BA_18m = (sum((pi*(DBH[real.dist <= 18]/20)^2)/10000)/(pi*18^2))*10000,
Broad_BA_17m = (sum((pi*(DBH[real.dist <= 17]/20)^2)/10000)/(pi*17^2))*10000,
Broad_BA_16m = (sum((pi*(DBH[real.dist <= 16]/20)^2)/10000)/(pi*16^2))*10000,
Broad_BA_15m = (sum((pi*(DBH[real.dist <= 15]/20)^2)/10000)/(pi*15^2))*10000,
Broad_BA_14m = (sum((pi*(DBH[real.dist <= 14]/20)^2)/10000)/(pi*14^2))*10000,
Broad_BA_13m = (sum((pi*(DBH[real.dist <= 13]/20)^2)/10000)/(pi*13^2))*10000,
Broad_BA_12m = (sum((pi*(DBH[real.dist <= 12]/20)^2)/10000)/(pi*12^2))*10000,
Broad_BA_11m = (sum((pi*(DBH[real.dist <= 11]/20)^2)/10000)/(pi*11^2))*10000,
Broad_BA_10m = (sum((pi*(DBH[real.dist <= 10]/20)^2)/10000)/(pi*10^2))*10000,
Broad_BA_09m = (sum((pi*(DBH[real.dist <= 9]/20)^2)/10000)/(pi*9^2))*10000,
Broad_BA_08m = (sum((pi*(DBH[real.dist <= 8]/20)^2)/10000)/(pi*8^2))*10000,
Broad_BA_07m = (sum((pi*(DBH[real.dist <= 7]/20)^2)/10000)/(pi*7^2))*10000,
Broad_BA_06m = (sum((pi*(DBH[real.dist <= 6]/20)^2)/10000)/(pi*6^2))*10000,
Broad_BA_05m = (sum((pi*(DBH[real.dist <= 5]/20)^2)/10000)/(pi*5^2))*10000,
Broad_BA_04m = (sum((pi*(DBH[real.dist <= 4]/20)^2)/10000)/(pi*4^2))*10000)



tree.data.plot3 <- broad.leaved %>%
  filter(!(Site %in% "Arville")) %>%
  group_by(Site, Plot) %>%
  summarise(
Broad_BA_10m = (sum((pi*(DBH[real.dist <= 10]/20)^2)/10000)/(pi*10^2))*10000,
Broad_BA_09m = (sum((pi*(DBH[real.dist <= 9]/20)^2)/10000)/(pi*9^2))*10000,
Broad_BA_08m = (sum((pi*(DBH[real.dist <= 8]/20)^2)/10000)/(pi*8^2))*10000,
Broad_BA_07m = (sum((pi*(DBH[real.dist <= 7]/20)^2)/10000)/(pi*7^2))*10000,
Broad_BA_06m = (sum((pi*(DBH[real.dist <= 6]/20)^2)/10000)/(pi*6^2))*10000,
Broad_BA_05m = (sum((pi*(DBH[real.dist <= 5]/20)^2)/10000)/(pi*5^2))*10000,
Broad_BA_04m = (sum((pi*(DBH[real.dist <= 4]/20)^2)/10000)/(pi*4^2))*10000)

# Fill in 10-20 m radi for sites other than Arville with 10m radius data

tree.data.plot3$Broad_BA_20m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_19m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_18m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_17m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_16m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_15m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_14m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_13m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_12m <- tree.data.plot3$Broad_BA_10m

tree.data.plot3$Broad_BA_11m <- tree.data.plot3$Broad_BA_10m


broad.ba2 <- tree.data.plot2[,which(names(tree.data.plot2) %in% c("Site", "Plot", "Broad_BA_20m", "Broad_BA_19m", "Broad_BA_18m", "Broad_BA_17m", "Broad_BA_16m", "Broad_BA_15m", "Broad_BA_14m", "Broad_BA_13m", "Broad_BA_12m", "Broad_BA_11m", "Broad_BA_10m", "Broad_BA_09m", "Broad_BA_08m", "Broad_BA_07m", "Broad_BA_06m", "Broad_BA_05m", "Broad_BA_04m"))]

broad.ba3 <- tree.data.plot3[,which(names(tree.data.plot3) %in% c("Site", "Plot", "Broad_BA_20m", "Broad_BA_19m", "Broad_BA_18m", "Broad_BA_17m", "Broad_BA_16m", "Broad_BA_15m", "Broad_BA_14m", "Broad_BA_13m", "Broad_BA_12m", "Broad_BA_11m", "Broad_BA_10m", "Broad_BA_09m", "Broad_BA_08m", "Broad_BA_07m", "Broad_BA_06m", "Broad_BA_05m", "Broad_BA_04m"))]

broad.ba <- rbind(broad.ba2, broad.ba3)

tree.data.plot <- tree.data.plot %>% plyr::join(broad.ba, by=c("Site","Plot"))

rm(tree.data.plot2)
rm(tree.data.plot3)
rm(broad.leaved)
rm(broad.ba)
rm(broad.ba2)
rm(broad.ba3)

tree.data.plot[is.na(tree.data.plot)] <- 0 #no trees = 0
```

Calculate the % of broadleaved areas on the basal area at different radi:

```{r}
tree.data.plot$perc_broad_20m <- tree.data.plot$Broad_BA_20m/tree.data.plot$Basal_Area_20m*100
tree.data.plot$perc_broad_19m <- tree.data.plot$Broad_BA_19m/tree.data.plot$Basal_Area_19m*100
tree.data.plot$perc_broad_18m <- tree.data.plot$Broad_BA_18m/tree.data.plot$Basal_Area_18m*100
tree.data.plot$perc_broad_17m <- tree.data.plot$Broad_BA_17m/tree.data.plot$Basal_Area_17m*100
tree.data.plot$perc_broad_16m <- tree.data.plot$Broad_BA_16m/tree.data.plot$Basal_Area_16m*100
tree.data.plot$perc_broad_15m <- tree.data.plot$Broad_BA_15m/tree.data.plot$Basal_Area_15m*100
tree.data.plot$perc_broad_14m <- tree.data.plot$Broad_BA_14m/tree.data.plot$Basal_Area_14m*100
tree.data.plot$perc_broad_13m <- tree.data.plot$Broad_BA_13m/tree.data.plot$Basal_Area_13m*100
tree.data.plot$perc_broad_12m <- tree.data.plot$Broad_BA_12m/tree.data.plot$Basal_Area_12m*100
tree.data.plot$perc_broad_11m <- tree.data.plot$Broad_BA_11m/tree.data.plot$Basal_Area_11m*100
tree.data.plot$perc_broad_10m <- tree.data.plot$Broad_BA_10m/tree.data.plot$Basal_Area_10m*100
tree.data.plot$perc_broad_09m <- tree.data.plot$Broad_BA_09m/tree.data.plot$Basal_Area_09m*100
tree.data.plot$perc_broad_08m <- tree.data.plot$Broad_BA_08m/tree.data.plot$Basal_Area_08m*100
tree.data.plot$perc_broad_07m <- tree.data.plot$Broad_BA_07m/tree.data.plot$Basal_Area_07m*100
tree.data.plot$perc_broad_06m <- tree.data.plot$Broad_BA_06m/tree.data.plot$Basal_Area_06m*100
tree.data.plot$perc_broad_05m <- tree.data.plot$Broad_BA_05m/tree.data.plot$Basal_Area_05m*100
tree.data.plot$perc_broad_04m <- tree.data.plot$Broad_BA_04m/tree.data.plot$Basal_Area_04m*100

#NaNs appear when dividing by 0, which happens in places with no trees (basal area = 0). Therefore, we can just replace NaNs by 0. In R, is.nan() does not work for dataset, so we will create a small function to address this.

is.nan.data.frame <- function(x)
do.call(cbind, lapply(x, is.nan))

tree.data.plot[is.nan(tree.data.plot)] <- 0 
```


## Hemispherical pictures

Canopy opennes was calculated with the script _HemiphotBatch.R_ separately for each site. This script comes from ter Steege, H. (2018).

```{r}
file.list <- list.files(path = "J:/HEMISPHERICAL DATA/", 
                        pattern = "*.csv", all.files = FALSE, full.names = TRUE, recursive = FALSE,
                        ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE)

df.list <- lapply(file.list, read.csv, sep = ";", dec = ",")

rm(file.list)
hemispherical.data <- do.call(rbind, df.list)
rm(df.list)
```

Keep only data of interest:

```{r}
hemispherical.data <- hemispherical.data[,which(colnames(hemispherical.data) %in% c("Site", "Plot", "CanOpen"))]
```

Calculate differences in summer and winter canopy openness for Siljan (the only site with both types of pictures). Data labelled as "Siljan" corresponds to summer season pictures (since all other sites only have pictures for summer). Data appearing as "Siljan_win" corresponds to winter pictures.

```{r}
hemispherical.data[which(hemispherical.data$Site == "Siljan_winter"),]$CanOpen -
hemispherical.data[which(hemispherical.data$Site == "Siljan"),]$CanOpen
```

For now, we will ignore this canopy cover differences and focus only on summer pictures:

```{r}
hemispherical.data <- hemispherical.data[-which(hemispherical.data$Site == "Siljan_winter"),]
```

Merge with main dataset:

```{r}
tree.data.plot <- tree.data.plot %>% plyr::join(hemispherical.data, by=c("Site","Plot"))

rm(hemispherical.data)
```

Export dataset:

```{r}
tree.data.plot <- tree.data.plot %>% 
  dplyr::rename(
    site = Site)

write.csv(tree.data.plot, file = "./DATASETS/tree_data.csv", row.names = FALSE, quote = FALSE)
```


# Figure

```{r}
tree.data.plot$site_label <- ifelse(tree.data.plot$site == "Arville", "B (50° N)", ifelse(tree.data.plot$site == "Hochwald", "G (49° N)", ifelse(tree.data.plot$site == "Vivarp", "SS (56° N)", ifelse(tree.data.plot$site == "Siljan", "MS (60° N)", "NS (64° N)"))))

tree.data.plot %>%
# filter(!(site %in% "Arville")) %>%
 ggplot() +
 aes(x = Basal_Area_20m,
     y = perc_broad_20m, colour = site_label) +
 geom_point(aes(shape = site_label), 
 size = 2.5) +
  scale_shape_manual(name = "Site",
                     values=c(15, 16, 17, 18, 20),
                     limits=c("NS (64° N)", "MS (60° N)", "SS (56° N)", "B (50° N)", "G (49° N)")) +
  scale_color_manual(name = "Site",
                     values=c("#E09F3E", "#89BDB9", "#335C67", "#65875A", "#9E2A2B"),
                     limits=c("NS (64° N)", "MS (60° N)", "SS (56° N)", "B (50° N)", "G (49° N)")) +
 labs(x = expression(paste("Basal Area (", m^2, " \U00B7 ", ha^-1,")")), y = "Broadleaves (%)", fill = "Site") +
 theme_minimal() +
 theme(axis.title.y = element_text(size = 16L), axis.title.x = element_text(size = 16L),
       legend.title = element_blank(),
       legend.text = element_text(size = 10),
       legend.position = "top")
#rm(list = ls())
```
