==> 5. Offset_calculation.Rmd <==
# Offset calculation and dataset merging

*Code reproducibility*

To allow a better reproducibility of this script, we recommend to use checkpoint() set at 2022-11-21. This allows to install and use the packages that were available and run at the time of making this script.

```{r}
#install.packages("checkpoint")
library(checkpoint)
checkpoint("2022-11-21")
```

Load packages...

```{r}
library(dplyr)
library(lubridate)
library(ggplot2)
library(reshape)
library(ggpubr)
```

First, we need to load all data:

```{r}
logger.temp <- read.csv("./DATASETS/logger_temp.csv", header = TRUE)

ERA5.temp <- read.csv("./DATASETS/ERA5_temp.csv", header = TRUE)

mcera5.temp <- read.csv("./DATASETS/mcera5.csv", header = TRUE)

weatherstation.temp <- read.csv("./DATASETS/weatherstations_merged.csv", header = TRUE)
```

And we merge everything (weather station data, ERA5, mcera5 and our loggers) by site and date:

```{r}
temp <- merge(merge(merge(logger.temp, ERA5.temp, by = c("site", "Date"), all.x = TRUE), weatherstation.temp, by = c("site", "Date"), all.x = TRUE), mcera5.temp, by = c("site", "Date"), all.x = TRUE)

rm(logger.temp)
rm(ERA5.temp)
rm(weatherstation.temp)
rm(mcera5.temp)
```

Calculate temperature offsets ERA5:

```{r}
temp$air_meanoff_ERA5 <- temp$Temp.mean - temp$era5_meanair # air mean temperature

temp[which(temp$location == "s"),]$air_meanoff_ERA5 <- NA # knock out values for soil loggers, as this will be calculated with "era5_soil"

temp$air_minoff_ERA5 <- temp$Temp.min - temp$era5_minair

temp[which(temp$location == "s"),]$air_minoff_ERA5 <- NA # remove minimum values for soil temperature, as offsets are calculated with minimum air temperature.

temp$air_maxoff_ERA5 <- temp$Temp.max - temp$era5_maxair

temp[which(temp$location == "s"),]$air_maxoff_ERA5 <- NA # remove maximum values for soil temperature, as offsets are calculated with minimum air

temp$soil_meanoff_ERA5 <- temp$Temp.mean-temp$era5_meansoil

temp[which(temp$location == "a"),]$soil_meanoff_ERA5 <- NA # remove values obtained for air loggers, as offsets have been calculated with soil temperature
```

Offset calculation for mcera5 data:

```{r}
temp$air_meanoff_mcera5 <- temp$Temp.mean - temp$meanair_mcera5 # air mean temperature

temp[which(temp$location == "s"),]$air_meanoff_mcera5 <- NA # knock out values for soil loggers, as this will be calculated with "era5_soil"

temp$air_minoff_mcera5 <- temp$Temp.min - temp$minair_mcera5

temp[which(temp$location == "s"),]$air_minoff_mcera5 <- NA # remove minimum values for soil temperature, as offsets are calculated with minimum air temperature.

temp$air_maxoff_mcera5 <- temp$Temp.max - temp$maxair_mcera5

temp[which(temp$location == "s"),]$air_maxoff_mcera5 <- NA # remove maximum values for soil temperature, as offsets are calculated with minimum air
```

Offset claculation weather stations (ws)

```{r}
temp$air_meanoff_ws <- temp$Temp.mean - temp$station_meanair # air mean temperature

temp[which(temp$location == "s"),]$air_meanoff_ws <- NA # knock out values for soil loggers, as this will be calculated with "era5_soil"

temp$air_minoff_ws <- temp$Temp.min - temp$station_minair

temp[which(temp$location == "s"),]$air_minoff_ws <- NA # remove minimum values for soil temperature, as offsets are calculated with minimum air temperature.

temp$air_maxoff_ws <- temp$Temp.max - temp$station_maxair

temp[which(temp$location == "s"),]$air_maxoff_ws <- NA # remove maximum values for soil temperature, as offsets are calculated with minimum air
```

Compare ait temperature offsets:

```{r}
offsetplot <- temp[which(temp$location == "a"),]

offsetplot$month <- month(offsetplot$Date)

offsetplot.jan <- offsetplot[which(offsetplot$month == 1),]

offsetplot.jan <- offsetplot.jan[,which(names(offsetplot.jan) %in% c("air_meanoff_ERA5", "air_minoff_ERA5", "air_maxoff_ERA5", "air_meanoff_mcera5", "air_minoff_mcera5", "air_maxoff_mcera5", "air_meanoff_ws", "air_minoff_ws", "air_maxoff_ws"))]

offsetplot.jul <- offsetplot[which(offsetplot$month == 7),]
rm(offsetplot)

offsetplot.jul <- offsetplot.jul[,which(names(offsetplot.jul) %in% c("air_meanoff_ERA5", "air_minoff_ERA5", "air_maxoff_ERA5", "air_meanoff_mcera5", "air_minoff_mcera5", "air_maxoff_mcera5", "air_meanoff_ws", "air_minoff_ws", "air_maxoff_ws"))]

melt.jan <- melt(offsetplot.jan)
rm(offsetplot.jan)

melt.jan$tempclass <- ifelse(grepl("max", melt.jan$variable, fixed = TRUE), "max", ifelse(grepl("min", melt.jan$variable, fixed = TRUE), "min", ifelse(grepl("mean", melt.jan$variable, fixed = TRUE), "mean", "error")))

melt.jan$sourceclass <- ifelse(grepl("_mcera5", melt.jan$variable, fixed = TRUE), "mcera5", ifelse(grepl("_ERA5", melt.jan$variable, fixed = TRUE), "ERA5", ifelse(grepl("_ws", melt.jan$variable, fixed = TRUE), "WS", "error")))

melt.jul <- melt(offsetplot.jul)
rm(offsetplot.jul)

melt.jul$tempclass <- ifelse(grepl("max", melt.jul$variable, fixed = TRUE), "max", ifelse(grepl("min", melt.jul$variable, fixed = TRUE), "min", ifelse(grepl("mean", melt.jul$variable, fixed = TRUE), "mean", "error")))

melt.jul$sourceclass <- ifelse(grepl("_mcera5", melt.jul$variable, fixed = TRUE), "mcera5", ifelse(grepl("_ERA5", melt.jul$variable, fixed = TRUE), "ERA5", ifelse(grepl("_ws", melt.jul$variable, fixed = TRUE), "WS", "error")))


# coldest month:

coldoffsetplot <-
  ggplot(melt.jan) +
  aes(x = value, y = variable, fill = sourceclass) +
  geom_violin() +
      geom_vline(xintercept=c(-0,0), linetype="dotted", col = "red", lwd = 1) +
  scale_fill_manual(
    values = c(ERA5 = "#e09f3e",
    mcera5 = "#9e2a2b",
    WS = "#335c67"),
    name = "Data source:",
    breaks = c("ERA5", "mcera5", "WS"),
    labels = c("ERA5 raw", "ERA5 adjusted [mcera5]", "Weather station")) +
  theme_minimal() + 
  facet_grid(tempclass ~ . , 
             scales = "free_y", 
             switch = "y",                
             labeller = as_labeller(
                 c(max = "Max offset", 
                   mean = "Mean offset", 
                   min = "Min offset"))) +
  labs(title = "Coldest month") +
  ylab(NULL) + 
  xlab("°C") + 
  theme(axis.text.y = element_blank(),
        plot.title = element_text(size = 22L,
                                  face = "bold",
                                  hjust = 0.5),
        strip.text.y = element_text(size = 18),
        axis.text.x = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        legend.title = element_text(size=18),
        legend.text = element_text(size=15)) +
  guides(fill = guide_legend(reverse=TRUE)) +
  geom_boxplot(width=0.1, fill = "white")


# warmest month:

warmoffsetplot <-
  ggplot(melt.jul) +
  aes(x = value, y = variable, fill = sourceclass) +
  geom_violin() +
      geom_vline(xintercept=c(-0,0), linetype="dotted", col = "red", lwd = 1) +
  scale_fill_manual(
    values = c(ERA5 = "#e09f3e",
    mcera5 = "#9e2a2b",
    WS = "#335c67"),
    name = "Data source:",
    breaks = c("ERA5", "mcera5", "WS"),
    labels = c("ERA5 raw", "ERA5 adjusted [mcera5]", "Weather station")) +
  theme_minimal() + 
  facet_grid(tempclass ~ . , 
             scales = "free_y", 
             switch = "y",                
             labeller = as_labeller(
                 c(max = "Max offset", 
                   mean = "Mean offset", 
                   min = "Min offset"))) +
  labs(title = "Warmest month") +
  ylab(NULL) + 
  xlab("°C") + 
  theme(axis.text.y = element_blank(),
        plot.title = element_text(size = 22L,
                                  face = "bold",
                                  hjust = 0.5),
        strip.text.y = element_text(size = 18),
        axis.text.x = element_text(size = 14),
        axis.title.x = element_text(size = 18),
        legend.title = element_text(size=18),
        legend.text = element_text(size=15)) +
  guides(fill = guide_legend(reverse=TRUE)) +
  geom_boxplot(width=0.1, fill = "white")

ggarrange(coldoffsetplot, warmoffsetplot, common.legend = TRUE, legend = "bottom")

#rm(coldoffsetplot)
#rm(warmoffsetplot)
#rm(melt.jan)
#rm(melt.jul)
```

Are any statistically significant differences across the offsets that are calculated with different macroclimatic sources?

```{r}
# coldest month

jan.meanoff <- melt.jan[which(melt.jan$tempclass == "mean"),]

ks.test(jan.meanoff$value, pnorm)

kruskal.test(value~sourceclass, data = jan.meanoff)
library(FSA)
dunnTest(value~sourceclass, data = jan.meanoff)


jan.minoff <- melt.jan[which(melt.jan$tempclass == "min"),]

ks.test(jan.minoff$value, pnorm)

kruskal.test(value~sourceclass, data = jan.minoff)
dunnTest(value~sourceclass, data = jan.minoff)

jan.maxoff <- melt.jan[which(melt.jan$tempclass == "max"),]
ks.test(jan.maxoff$value, pnorm)

kruskal.test(value~sourceclass, data = jan.maxoff)
dunnTest(value~sourceclass, data = jan.maxoff)

# warmest month

jul.meanoff <- melt.jul[which(melt.jul$tempclass == "mean"),]

ks.test(jul.meanoff$value, pnorm)

kruskal.test(value~sourceclass, data = jul.meanoff)
dunnTest(value~sourceclass, data = jul.meanoff)


jul.minoff <- melt.jul[which(melt.jul$tempclass == "min"),]

ks.test(jul.minoff$value, pnorm)

kruskal.test(value~sourceclass, data = jul.minoff)
dunnTest(value~sourceclass, data = jul.minoff)

jul.maxoff <- melt.jul[which(melt.jul$tempclass == "max"),]
ks.test(jul.maxoff$value, pnorm)

kruskal.test(value~sourceclass, data = jul.maxoff)
dunnTest(value~sourceclass, data = jul.maxoff)
```


# Merge with tree data

Load tree data:

```{r}
tree.data <- read.csv("./DATASETS/tree_data.csv", header = TRUE)
```

Merge the temperature dataset and the tree data:

```{r}
tree.temp <- merge(temp, tree.data, by = c("site", "Plot"))

rm(temp)
rm(tree.data)
```

Export the merged dataset:

```{r}
write.csv(tree.temp, file = "./DATASETS/tree_temp.csv", row.names = FALSE, quote = FALSE)

rm(list = ls())
```
